+++ pkg/gpu/nvml/testutil/nvml.go
@@ -8,12 +8,14 @@
 package testutil
 
 import (
+	"fmt"
 	"testing"
 
 	"github.com/NVIDIA/go-nvml/pkg/nvml"
 	"github.com/stretchr/testify/require"
 
 	ddnvml "github.com/DataDog/datadog-agent/pkg/gpu/nvml"
+	"github.com/DataDog/datadog-agent/pkg/gpu/testutil"
 )
 
 // ToDDNVMLDevices converts a slice of nvml.Device to a slice of ddnvml.Device
@@ -39,8 +41,34 @@ func GetDDNVMLMocksWithIndexes(t testing.TB, indexes ...int) []*ddnvml.Device {
 // GetDDNVMLMockWithIndex returns a ddnvml.Device mock with the given index, based on the data
 // present in mocks.go
 func GetDDNVMLMockWithIndex(t testing.TB, index int) *ddnvml.Device {
-	dev := GetDeviceMock(index)
+	dev := testutil.GetDeviceMock(index)
 	dddev, err := ddnvml.NewDevice(dev)
 	require.NoError(t, err, "error converting nvml.Device to ddnvml.Device")
 	return dddev
+}
+
+// RequireDevicesEqual checks that the two devices are equal by comparing their UUIDs, which gives a better
+// output than using require.Equal on the devices themselves
+func RequireDevicesEqual(t *testing.T, expected, actual *ddnvml.Device, msgAndArgs ...interface{}) {
+	extraFmt := ""
+	if len(msgAndArgs) > 0 {
+		extraFmt = fmt.Sprintf(msgAndArgs[0].(string), msgAndArgs[1:]...) + ": "
+	}
+
+	require.Equal(t, expected.UUID, actual.UUID, "%sUUIDs do not match", extraFmt)
+}
+
+// RequireDeviceListsEqual checks that the two device lists are equal by comparing their UUIDs, which gives a better
+// output than using require.ElementsMatch on the lists themselves
+func RequireDeviceListsEqual(t *testing.T, expected, actual []*ddnvml.Device, msgAndArgs ...interface{}) {
+	extraFmt := ""
+	if len(msgAndArgs) > 0 {
+		extraFmt = fmt.Sprintf(msgAndArgs[0].(string), msgAndArgs[1:]...) + ": "
+	}
+
+	require.Len(t, actual, len(expected), "%sdevice lists have different lengths", extraFmt)
+
+	for i := range expected {
+		require.Equal(t, expected[i].UUID, actual[i].UUID, "%sUUIDs do not match for element %d", extraFmt, i)
+	}
 }
